backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
pydantic==2.6.3
python-multipart==0.0.9
pdfplumber==0.11.4
rapidfuzz==3.9.7
numpy==1.26.4
google-generativeai==0.8.5
python-dotenv==1.0.1
pytest==8.3.5
